pyproject.toml
[project]
name = "proxy"
version = "1.0.0"
description = "获取代理服务器"
readme = "README.md"
requires-python = "==3.14.4"
dependencies = [
    "beautifulsoup4==4.14.3",
    "dnspython==2.8.0",
    "fastapi[standard]==0.135.3",
    "pyyaml==6.0.3",
    "requests==2.33.1",
]

[dependency-groups]
dev = [
    "python-dotenv==1.2.2",
    "ruff==0.15.10",
    "ty==0.0.29",
    "types-requests==2.33.0.20260408",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/proxy"]

[tool.ruff]
line-length = 88
target-version = "py314"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle (error)
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "B",   # bugbear
    "C",   # flake8-comprehensions
    "SIM", # flake8-simplify
    "I",   # isort
    "UP",  # pyupgrade
    "PIE", # flake8-pie
    "PGH", # pygrep-hooks
    "PYI", # flake8-pyi
    "RUF",
    "S",   # flake8-bandit
]

ignore = ["E501", "S311"]

[tool.ruff.lint.isort]
combine-as-imports = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

[tool.ty.environment]
root = ["./src"]
